requirements.txt
numpy
cupy-cuda12x
jax[cuda12]
warp-lang
matplotlib
tqdm
scipy
pytest
taichi